Allow overriding the ServiceAccountIssuer for IRSA
