+++ pkg/model/components/BUILD.bazel
@@ -32,8 +32,8 @@ go_library(
         "//pkg/apis/kops:go_default_library",
         "//pkg/apis/kops/util:go_default_library",
         "//pkg/assets:go_default_library",
+        "//pkg/dns:go_default_library",
         "//pkg/k8sversion:go_default_library",
-        "//pkg/model/iam:go_default_library",
         "//pkg/wellknownports:go_default_library",
         "//upup/pkg/fi:go_default_library",
         "//upup/pkg/fi/cloudup/gce:go_default_library",
