+++ upup/pkg/fi/cloudup/tests/bootstrapchannelbuilder/service-account-iam/manifest.yaml
@@ -6,7 +6,7 @@ spec:
   addons:
   - id: k8s-1.16
     manifest: kops-controller.addons.k8s.io/k8s-1.16.yaml
-    manifestHash: 40ffaead818a21b374588ed2f94b517b9569c71f
+    manifestHash: 90b1206425c7a034147e2373ea00193018a1fb4a
     name: kops-controller.addons.k8s.io
     needsRollingUpdate: control-plane
     selector:
@@ -17,17 +17,11 @@ spec:
     selector:
       k8s-addon: core.addons.k8s.io
   - id: k8s-1.12
-    manifest: kube-dns.addons.k8s.io/k8s-1.12.yaml
-    manifestHash: 470684b80af41aa53c25bc5ac1a78b259c93336b
-    name: kube-dns.addons.k8s.io
+    manifest: coredns.addons.k8s.io/k8s-1.12.yaml
+    manifestHash: 004bda4e250d9cec5d5f3e732056020b78b0ab88
+    name: coredns.addons.k8s.io
     selector:
-      k8s-addon: kube-dns.addons.k8s.io
-  - id: k8s-1.8
-    manifest: rbac.addons.k8s.io/k8s-1.8.yaml
-    manifestHash: 38ba4e0078bb1225421114f76f121c2277ca92ac
-    name: rbac.addons.k8s.io
-    selector:
-      k8s-addon: rbac.addons.k8s.io
+      k8s-addon: coredns.addons.k8s.io
   - id: k8s-1.9
     manifest: kubelet-api.rbac.addons.k8s.io/k8s-1.9.yaml
     manifestHash: 8ee090e41be5e8bcd29ee799b1608edcd2dd8b65
